Matrix task v 1.7

+++ tasks/seq/sadikov_I_sum_values_by_columns_matrix/include/sq_task.h
@@ -10,13 +10,13 @@
 #include "core/task/include/task.hpp"
 // commit
 namespace sadikov_I_Sum_values_by_columns_matrix_seq {
-std::shared_ptr<ppc::core::TaskData> CreateTaskData(std::vector<double> &InV, const std::vector<size_t> &CeV,
-                                                    std::vector<double> &OtV);
+std::shared_ptr<ppc::core::TaskData> CreateTaskData(std::vector<int> &InV, const std::vector<size_t> &CeV,
+                                                    std::vector<int> &OtV);
 std::vector<double> Randvector(size_t size);
 class MatrixTask : public ppc::core::Task {
  private:
-  std::vector<double> sum;
-  std::vector<double> matrix;
+  std::vector<int> sum;
+  std::vector<int> matrix;
   size_t rows_count, columns_count;
 
  public:
